requirements.txt
pandas
numpy
bs4
lmxl
ffmpeg-python
openai-whisper
transformers

--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchaudio

# torchvision --extra-index-url https://download.pytorch.org/whl/cu116
# torchtext --extra-index-url https://download.pytorch.org/whl/cu116
# torch --extra-index-url https://download.pytorch.org/whl/cu116
# torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu116